categories connection in usermode between client and server and connection between server and client in books in admin

+++ client/src/app/services/books.service.ts
@@ -18,15 +18,15 @@ export class BooksService {
     console.log(bookData);
     return this._HttpClient.post('http://localhost:3000/admin/books',bookData)
   }
-  getAllBooks(): Observable<any> {
-    return this._HttpClient.get('http://localhost:3000/books');
+  getAllBooks(page:number,limit:number): Observable<any> {
+    return this._HttpClient.get(`http://localhost:3000/admin/books?page=${page}&limit=${limit}`);
   }
 
   deleteBookById(id: number): Observable<any> {
-    return this._HttpClient.delete(`http://localhost:3000/books/${id}`);
+    return this._HttpClient.delete(`http://localhost:3000/admin/books/${id}`);
   }
 
   editBook(id:number,data:object){
-   return this._HttpClient.patch(`http://localhost:3000/books/${id}`,data)
+   return this._HttpClient.patch(`http://localhost:3000/admin/books/${id}`,data)
   }
 }
